Fea/user/policies

+++ app/Http/Controllers/V1/Api/Admin/UserController.php
@@ -16,8 +16,9 @@ class UserController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
+        $this->authorize('index', $request->user());
         $user = User::where('status', '1')->paginate(10);
 
         return response()->json([
@@ -43,6 +44,7 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('create', $request->user());
         $user = new User();
         $user->fill($request->all());
         if ($user->save()) {
@@ -63,8 +65,9 @@ class UserController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function show(User $user)
+    public function show(Request $request, User $user)
     {
+        $this->authorize('view', $request->user(), $user);
         return response()->json([
             new UserResource($user),
         ], 200);
@@ -91,6 +94,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('update', $request->user(), $user);
         $user->update($request->all());
         return response()->json([
             'message' => "Successfully Updated User"
@@ -103,10 +107,10 @@ class UserController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy(Request $request, User $user)
     {
+        $this->authorize('delete', $request->user());
         $result = $user->delete();
-
         if ($result) {
             return response()->json('', 204);
         }
